Fix implicit conversion warnings in ExPolygonsIndex.cpp

+++ src/libslic3r/ExPolygonsIndex.cpp
@@ -39,13 +39,13 @@ ExPolygonsIndex ExPolygonsIndices::cvt(uint32_t index) const
     // find expolygon index
     auto fn = [](const std::vector<uint32_t> &offsets, uint32_t index) { return offsets[0] < index; };
     auto it = std::lower_bound(m_offsets.begin() + 1, m_offsets.end(), index, fn);
-    result.expolygons_index = it - m_offsets.begin();
+    result.expolygons_index = static_cast<uint32_t>(it - m_offsets.begin());
     if (it == m_offsets.end() || it->at(0) != index) --result.expolygons_index;
 
     // find polygon index
     const std::vector<uint32_t> &shape_offset = m_offsets[result.expolygons_index];
     auto it2 = std::lower_bound(shape_offset.begin() + 1, shape_offset.end(), index);
-    result.polygon_index = it2 - shape_offset.begin();
+    result.polygon_index = static_cast<uint32_t>(it2 - shape_offset.begin());
     if (it2 == shape_offset.end() || *it2 != index) --result.polygon_index;
 
     // calculate point index
